Add files via upload
Git Bash was not agreeing with my laptop, I believe it's a PATH issue that I have to correct but I am choosing to do it in Git Hub because it is an open resource to us and because I know the basics of the platform and want to have at least something to turn in. I attempted GIT in the command line where I had the project files initially. I did this in case I can't figure out how to fix my path, I apologize for the inconvenience.